Fix incorrect signal handler loop (#5)
Issue: The loop to close the events in the signal handling thread failed to decrement
the loop counter.

Impact: Minor. This would cause the signal handling thread to spin closing handles,
however the code is only executed if an Event wait fails and the thread attempts to
terminate which is an abnormal condition.

Fix: Rework the loop to correctly decrement the index.
Closes #5

Signed-off-by: Nick Connolly <47530082+nconnolly1@users.noreply.github.com>